LPS-58904 move panel appps to the right modules
