+++ modules/apps/workflow/workflow-task-web/build.gradle
@@ -1,5 +1,6 @@
 dependencies {
 	compile group: "org.osgi", name: "org.osgi.compendium", version: "5.0.0"
+	compile project(":apps:application-list:application-list-api")
 }
 
 liferay {
